# documentation/openapi.yml
openapi : 3.0.3
info :
    description : This is the specification for the Stadtgemüse Service.
    version : 2.0.0
    title : Stadtgemüse Service
    contact : { }
servers :
    -   url : 'https://veg-service.herokuapp.com/v2/'
        description : Production v2
...
paths :
    '/balance' :
        get :
            security :
                -   bearerAuth : [ ]
            tags :
                - Balance
            summary : Gets the current balance of the user
            description : 'Gets the current balance of the user, or 0 if the user is currently not known'
            operationId : balanceGet
            responses :
                '200' :
                    description : The current value of the balance for the specified user
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/balance-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
        patch :
            security :
                -   bearerAuth : [ ]
            tags :
                - Balance
            summary : Changes the balance for the specified user
            operationId : balancePatch
            requestBody :
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/balance-patch-request'
            responses :
                '200' :
                    description : The current value of the balance for the specified user
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/balance-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
    '/balance/topup' :
        post :
            security :
                -   bearerAuth : [ ]
            tags :
                - Balance
            summary : Topup the balance for the specified user
            operationId : balanceTopup
            requestBody :
                description : 'Top up the balance with the amount, containing a positive value'
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/balance-amount-request'
            responses :
                '200' :
                    description : The current value of the balance for the specified user
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/balance-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
    '/balance/withdraw' :
        post :
            security :
                -   bearerAuth : [ ]
            tags :
                - Balance
            summary : Withdraw the balance for the specified user
            operationId : balanceWithdraw
            requestBody :
                description : 'Withdraw the amount from the balance, containing a positive value'
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/balance-amount-request'
            responses :
                '200' :
                    description : The current value of the balance for the specified user
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/balance-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
...
    /user :
        get :
            security :
                -   bearerAuth : [ ]
            tags :
                - User
            summary : Get own user data
            operationId : userGet
            responses :
                '200' :
                    description : Get own user data
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/user-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
        delete :
            security :
                -   bearerAuth : [ ]
            tags :
                - User
            summary : Delete user account
            operationId : userDelete
            responses :
                '204' :
                    description : The user account was deleted
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
...
    /auth/refresh :
        post :
            tags :
                - Authentication
            summary : Get a new set of tokens using the refresh token
            description : |-
                Get a new set of tokens using the refresh token. The refreshToken used will be revoked, thus subsequent
                calls with this token will fail.
            operationId : refresh
            requestBody :
                description : The body containing the refresh token
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/refresh-request'
            responses :
                '201' :
                    description : Token succesfully retrieved
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/token-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
        delete :
            security :
                -   bearerAuth : [ ]
            tags :
                - Authentication
            summary : Revoke a refresh token
            operationId : revoke
            requestBody :
                description : The body containing the refresh token to revoke
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/refresh-request'
            responses :
                '204' :
                    description : Token succesfully revoked
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
    /auth/refresh/all :
        delete :
            security :
                -   bearerAuth : [ ]
            tags :
                - Authentication
            summary : Revoke all refresh tokens for a user
            operationId : revokeAll
            responses :
                '204' :
                    description : Tokens succesfully revoked
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
components :
    securitySchemes :
        bearerAuth : # arbitrary name for the security scheme
            type : http
            scheme : bearer
            bearerFormat : JWT
    schemas :
...
        error-response :
            description : "Contains details about the occurred error"
            type : object
            required :
                - errorCode
                - status
                - errorMessage
            properties :
                errorCode :
                    description : "Number, that can be referenced to a specific error"
                    example : 400001
                    type : integer
                status :
                    description : "HTTP Status Code returned"
                    example : 400
                    type : integer
                message :
                    description : "A detailed error message"
                    example : "Validation failed"
                    type : string
                details :
                    type : array
                    items :
                        $ref : '#/components/schemas/error-detail'
        error-detail :
            description : A more precise explaination of an error
            type : object
            properties :
                message :
                    description : A message relating to the error detail
                    type : string
                    example : "Invalid request parameter: balance"
                detail :
                    description : A detailed message of the error
                    type : string
                    example : "The value null does not match the Specification"

        #
        # Request Schemas
        #
        balance-patch-request :
            type : object
            description : Contains data about how to change the balance
            required :
                - balance
            properties :
                balance :
                    description : Absolute balance to be applied instead of the current balance
                    example : 60
                    type : number
                    format : float
        balance-amount-request :
            description : Contains data about the amount you want to change the balance
            type : object
            required :
                - amount
            properties :
                amount :
                    description : Realative amount to be applied to the balance
                    example : 60
                    minimum : 0
                    type : number
                    format : float
...
        refresh-request :
            type : object
            description : Object containing the refresh token
            required :
                - refreshToken
            properties :
                refreshToken :
                    $ref : '#/components/schemas/token'
...
        balance-response :
            description : Contains data about the current balance
            type : object
            required :
                - balance
                - name
            properties :
                balance :
                    description : The current balance
                    example : 10
                    type : number
                    format : float
                name :
                    description : The name of the owner of this balance
                    example : Robby
                    type : string
...
        user-response :
            type : object
            properties :
                id :
                    $ref : '#/components/schemas/uuid'
                username :
                    $ref : '#/components/schemas/username'
                email :
                    $ref : '#/components/schemas/email'
                memberId :
                    $ref : '#/components/schemas/memberId'
        token-response :
            type : object
            properties :
                token :
                    $ref : '#/components/schemas/token'
                refreshToken :
                    $ref : '#/components/schemas/token'

        #
        # Reusable Items
        #
        uuid :
            description : Unique identifier
            example : "0794e2e2-f5b2-42fb-a2b3-e12c75897f85"
            type : string
...
        username :
            description : The unique name of the user
            type : string
            example : Robby
            maxLength : 255
...
        email :
            description : An e-mail address
            type : string
            format : email
            example : robby@stadtgemuese-wiesbaden.de
            maxLength : 255
        memberId :
            description : The stadtgemuese member id
            type : string
            example : 1234
            maxLength : 255
...
        token :
            description : A JWT token
            type : string
            example : eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
            maxLength : 10000
...